adds logistics map filters
