--- frontend/scripts/react-components/nav/filters-nav/burger.scss
@@ -1,77 +1,0 @@
-@import '../../../../styles/settings';
-
-.c-burger {
-  position: relative;
-  width: 18px;
-  height: $nav-height;
-
-  transform: rotate(0deg);
-  cursor: pointer;
-
-  &:hover > .ingredient {
-    background: $medium-pink;
-  }
-
-  > .ingredient {
-    display: block;
-    position: absolute;
-    left: 0;
-    width: 100%;
-    height: 2px;
-    transform: rotate(0deg);
-    border-radius: 3px;
-    background: $black;
-    opacity: 1;
-
-    &:first-child {
-      top: 26px;
-
-      transform-origin: left center;
-    }
-
-    &:nth-child(2) {
-      top: 31px;
-
-      transform-origin: left center;
-    }
-
-    &:last-child {
-      top: 36px;
-
-      transform-origin: left center;
-    }
-  }
-
-  &.open {
-
-    &:hover {
-      > .ingredient {
-        background: $medium-pink;
-      }
-    }
-
-    > .ingredient {
-      border-radius: 0;
-      background: $black;
-
-      &:first-child {
-        top: 25px;
-        left: 3px;
-
-        transform: rotate(45deg);
-      }
-
-      &:nth-child(2) {
-        width: 0%;
-        opacity: 0;
-      }
-
-      &:last-child {
-        top: 38px;
-        left: 3px;
-
-        transform: rotate(-45deg);
-      }
-    }
-  }
-}
