--- frontend/scripts/react-components/nav/filters-nav/filters-nav.scss
@@ -1,178 +1,0 @@
-@import 'styles/settings';
-
-.c-filters-nav {
-  display: flex;
-  top: 0;
-  position: fixed;
-  width: 100%;
-  height: $nav-height;
-  font-family: $font-family-1;
-  z-index: $z-nav;
-  background-color: $white;
-  box-shadow: 0 10px 20px 0 rgba($charcoal-grey, 0.2);
-
-  .c-dropdown .dropdown-list {
-    top: 100%;
-  }
-
-  .c-locale-selector {
-
-    .locale-selector-selected-item,
-    .locale-selector-menu-item {
-      color: $charcoal-grey;
-      font-family: $font-family-1;
-      font-size: 16px;
-      font-weight: $font-weight-bold;
-      letter-spacing: .5px;
-      text-align: center;
-      text-transform: uppercase;
-    }
-
-    .locale-selector-selected-item {
-      &::after {
-        border-top-color: $charcoal-grey;
-      }
-    }
-
-    .locale-selector-menu {
-      background-color: $white;
-      box-shadow: 0 3px 6px 0 rgba($black, 0.16), 0 3px 6px 0 rgba($black, 0.23);
-      top: 100%;
-      margin: 0;
-      width: 100%;
-
-      .locale-selector-menu-item {
-        width: 100%;
-        margin: 0;
-        padding: 10px;
-        opacity: 1;
-        font-weight: $font-weight-regular;
-
-        &:hover {
-          background-color: $manilla;
-        }
-
-        &:not(:first-child) {
-          border-top: solid 1px rgba($charcoal-grey, .2);
-        }
-      }
-    }
-  }
-
-  .icon-search {
-    width: 21px;
-    height: 22px;
-    fill: $charcoal-grey;
-  }
-}
-
-.filters-nav-section-container {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  height: 100%;
-  width: calc(100% - 55px);
-}
-
-.filters-nav-left-section,
-.filters-nav-right-section {
-  display: flex;
-  height: 100%;
-}
-
-.filters-nav-left-section {
-  border-right: 1px solid $pale-grey;
-}
-
-.filters-nav-right-section {
-  border-left: 1px solid $pale-grey;
-
-  .filters-nav-item:first-child {
-    border-left: 0;
-  }
-}
-
-
-
-.filters-nav-submenu-list {
-  display: flex;
-
-  .filters-nav-item {
-    padding: 0;
-  }
-
-  .filters-nav-link {
-    text-transform: uppercase;
-  }
-}
-
-.filters-nav-item {
-  display: inline-flex;
-  position: relative;
-  height: $nav-height;
-  align-items: center;
-  font-size: 16px;
-  font-weight: bold;
-  letter-spacing: .5px;
-  color: $charcoal-grey;
-  border-left: 1px solid $pale-grey;
-  cursor: pointer;
-
-  > * {
-    padding: 0 18px;
-  }
-
-
-  &.-no-padding {
-    > * {
-      padding: 0;
-      margin: 0 18px;
-    }
-  }
-
-  &.-icon {
-    > svg {
-      height: 21px;
-      width: 21px;
-    }
-  }
-
-  .dropdown-item-legend-summary {
-    padding: 0;
-  }
-}
-
-.filters-nav-link {
-  position: relative;
-  display: flex;
-  align-items: center;
-  height: 100%;
-  padding: 0 16px;
-  color: $charcoal-grey;
-
-  &::before {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    height: 6px;
-    width: 100%;
-    background-color: $medium-pink;
-    z-index: $z-base;
-  }
-
-  &.-active {
-    &::before {
-      content: '';
-    }
-  }
-
-  &:hover {
-    &::before {
-      content: '';
-    }
-  }
-}
-
-.filters-nav-item-logo {
-  cursor: pointer;
-}
