--- frontend/scripts/react-components/nav/filters-nav/nav-dropdown-selector/nav-dropdown-selector.js
@@ -1,1 +1,0 @@
-export { default } from './nav-dropdown-selector.component';
