--- frontend/scripts/react-components/nav/filters-nav/recolor-by-selector/recolor-by.component.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import RecolorBy from 'react-components/shared/recolor-by';
-
-function RecolorBySelector(props) {
-  return <RecolorBy {...props} size="rg" variant="nav" weight="regular" />;
-}
-
-export default RecolorBySelector;
