--- frontend/scripts/react-components/nav/filters-nav/tool-selection-modal-button/tool-selection-modal-button.scss
@@ -1,29 +1,0 @@
-@import 'styles/settings';
-@import 'styles/mixins';
-
-.c-tool-selection-modal-button {
-  height: 100%;
-  padding: 0 20px;
-  margin-right: 20px;
-  border-left: 1px solid $charcoal-grey-faded-a-lot;
-  position: relative;
-  cursor: pointer;
-  text-align: left;
-
-  &[disabled=""] {
-    cursor: default;
-  }
-
-  .selected-item {
-    display: block;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-
-    @include simple-arrow($charcoal-grey-faded);
-
-    &:after {
-      top: calc(50% + 10px);
-    }
-  }
-}
