+++ frontend/scripts/react-components/shared/sentence-selector/years-selector/years-selector.component.jsx
