+++ frontend/scripts/react-components/shared/sentence-selector/years-selector/years-selector.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { selectYears } from 'app/app.actions';
-import YearsSelector from 'react-components/nav/filters-nav/years-selector/years-selector.component';
+import YearsSelector from 'react-components/shared/sentence-selector/years-selector/years-selector.component';
 import { getToolYearsProps } from 'react-components/tool-links/tool-links.selectors';
 
 const mapStateToProps = state => getToolYearsProps(state);
