+++ frontend/scripts/react-components/shared/tool-bar/tool-bar.scss
@@ -73,7 +73,7 @@
 
       &.-no-padding {
         > * {
-          padding: 0;
+          padding: 0 0 20px 0;
         }
       }
     }
@@ -101,6 +101,14 @@
             > * {
               padding: 0 40px 0 24px;
             }
+          }
+        }
+      }
+
+      .slot-item {
+        &.-no-padding {
+          > * {
+            padding: 0 0 0 24px;
           }
         }
       }
